Read bible mode #149 Added new route
